% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{addChild}
\alias{addChild}
\title{Add a child at a point in the tree.}
\usage{
addChild(self, private, id, data)
}
\arguments{
\item{self}{The point in the tree where the child should be added.}

\item{private}{The private part of the tree.}

\item{id}{The id of the node that should be added.}

\item{data}{The data of the node that should be added.}
}
\value{
invisible the new node that was created.
}
\description{
Add a child at a point in the tree.
}
\keyword{internal}

