% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{addSibling}
\alias{addSibling}
\title{Add a sibling to the current node.}
\usage{
addSibling(self, private, data, id)
}
\arguments{
\item{self}{The point in the tree where the sibling should be added.}

\item{private}{The private part of the tree.}

\item{data}{The data of the node that should be added.}

\item{id}{The id of the node that should be added.}
}
\value{
invisible the new node that was created.
}
\description{
Add a sibling to the current node.
}
\keyword{internal}

