% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{addSiblingNode}
\alias{addSiblingNode}
\title{Add a node to the list of siblings of the current node.}
\usage{
addSiblingNode(self, private, node)
}
\arguments{
\item{self}{The point in the tree where the sibling should be added.}

\item{private}{The private part of the tree.}
}
\description{
Add a node to the list of siblings of the current node.
}
\keyword{internal}

