% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casting.R
\name{as.data.frame.GeneralTree}
\alias{as.data.frame.GeneralTree}
\title{Convert a GeneralTree to a data frame.}
\usage{
\method{as.data.frame}{GeneralTree}(x, row.names = NULL, optional = NULL,
  ...)
}
\arguments{
\item{x}{GeneralTree to convert to a data frame.}

\item{row.names}{Ignored.}

\item{optional}{Ignored.}

\item{...}{Ignored.}
}
\description{
Convert a GeneralTree to a data frame.
}

