% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casting.R
\name{as.GeneralTree.data.frame}
\alias{as.GeneralTree.data.frame}
\title{Convert a data frame to a GeneralTree.}
\usage{
\method{as.GeneralTree}{data.frame}(x, id = "id", data = "data",
  parent = "parent")
}
\description{
Convert a data frame to a GeneralTree.
}

