% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casting.R
\name{as.GeneralTree.expression}
\alias{as.GeneralTree.expression}
\title{Convert an object to a GeneralTree.}
\usage{
\method{as.GeneralTree}{expression}(x, ...)
}
\arguments{
\item{x}{The object that should be converted.}

\item{...}{passed to underlying functions.}
}
\description{
Convert an object to a GeneralTree.
}

