% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casting.R
\name{as.GeneralTree}
\alias{as.GeneralTree}
\title{Convert an object to a GeneralTree.}
\usage{
as.GeneralTree(x)
}
\arguments{
\item{x}{The object that should be converted.}
}
\description{
Convert an object to a GeneralTree.
}

