% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{<-.GeneralTree}
\alias{<-.GeneralTree}
\title{Deep clone a General Tree.}
\usage{
`<-.GeneralTree`(x, value)
}
\arguments{
\item{x}{The target to where the tree should be copied.}

\item{value}{The general tree that should be cloned into.}
}
\value{
a clone of the tree.
}
\description{
Deep clone a General Tree.
}

