% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmarkGeneralTree}
\alias{benchmarkGeneralTree}
\title{Benchmark function that measures the speed of the tree implementation.}
\usage{
benchmarkGeneralTree(depth = 4, number_of_splits = c(2, 4, 16),
  times = 100L)
}
\description{
Benchmark function that measures the speed of the tree implementation.
}

