% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{branch_depth}
\alias{branch_depth}
\title{Get the depth of a branch.}
\usage{
branch_depth(self, private)
}
\arguments{
\item{self}{The node where to start.}

\item{private}{The private members of a node.}
}
\value{
A numeric value that indicates the number of layers in the branch.
}
\description{
Get the depth of a branch.
}
\keyword{internal}

