% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{branchToList}
\alias{branchToList}
\title{Convert a branch to a list.}
\usage{
branchToList(self)
}
\arguments{
\item{self}{The branch where to start}
}
\value{
list that contains all the nodes in the branch.
}
\description{
Convert a branch to a list.
}
\keyword{internal}

