% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{data}
\alias{data}
\title{Return the data associated with a node.}
\usage{
data(self, private)
}
\arguments{
\item{self}{The node which data should be retrieved.}

\item{private}{The private members of a node.}
}
\value{
the data of the node.
}
\description{
Return the data associated with a node.
}
\keyword{internal}

