% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{deep_clone.GeneralTree}
\alias{deep_clone.GeneralTree}
\title{Deep clone a General Tree.}
\usage{
\method{deep_clone}{GeneralTree}(x)
}
\arguments{
\item{x}{The general tree that should be deep cloned.}
}
\value{
a clone of the tree.
}
\description{
Deep clone a General Tree.
}

