% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{deep_clone}
\alias{deep_clone}
\title{Deep clone a General tree}
\usage{
deep_clone(x)
}
\arguments{
\item{x}{The object that should be converted.}

\item{...}{passed to underlying functions.}
}
\description{
Deep clone a General tree
}

