% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{delete}
\alias{delete}
\title{Delete all a node and all nodes below that node.}
\usage{
delete(self, private)
}
\arguments{
\item{self}{the GeneralTree}

\item{private}{the private members of the GeneralTree.}
}
\description{
Delete all a node and all nodes below that node.
}

