% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{deleteId}
\alias{deleteId}
\title{Delete last referenced key.}
\usage{
deleteId(self, private, key)
}
\arguments{
\item{self}{The reference to the tree where the id should be searched.}

\item{private}{the private members of the GeneralTree.}

\item{id}{The id that should be deleted.}
}
\description{
Delete last referenced key.
}
\keyword{internal}

