% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{deleteKey}
\alias{deleteKey}
\title{Delete a node with a given id.}
\usage{
deleteKey(self, private, key)
}
\arguments{
\item{self}{The reference to the tree where the id should be searched.}

\item{id}{The id that should be deleted.}
}
\description{
Delete a node with a given id.
}
\keyword{internal}

