% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{depth}
\alias{depth}
\title{Returns the depth of the tree.}
\usage{
depth(self, private)
}
\arguments{
\item{self}{The reference to the tree where the id should be searched.}

\item{private}{the private members of the GeneralTree.}
}
\description{
Returns the depth of the tree.
}
\keyword{internal}

