% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{==.GeneralTree}
\alias{==}
\alias{==.GeneralTree}
\title{Compare two General Trees.}
\usage{
`==.GeneralTree`(x, value)
}
\arguments{
\item{x}{The target to where the tree should be copied.}

\item{value}{The general tree that should be cloned into.}
}
\value{
the result of the comparison.
}
\description{
Compare two General Trees.
}

