% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\docType{data}
\name{GeneralTree}
\alias{GeneralTree}
\title{GeneralTree is a R6 implementation of a tree that can have multiple child
per parent.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
GeneralTree
}
\description{
GeneralTree is a R6 implementation of a tree that can have multiple child
per parent.
}
\keyword{datasets}

