% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{generate_grViz}
\alias{generate_grViz}
\title{Create a DiagrammeR graph that represents the tree.}
\usage{
generate_grViz(obj, what = c("id", "data"), ...)
}
\arguments{
\item{what}{select what to draw in the tree.}

\item{...}{Additional arguments passed to create_nodes}
}
\description{
Create a DiagrammeR graph that represents the tree.
}
\seealso{
\code{\link{DiagrammeR::create_nodes}}
}

