% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{getBranchKeys}
\alias{getBranchKeys}
\title{Get the keys of the branch nodes below the current node.}
\usage{
getBranchKeys(self, private, key)
}
\arguments{
\item{self}{The node where to start.}

\item{private}{the private members of the GeneralTree.}
}
\value{
the keys associated with branch nodes.
}
\description{
Get the keys of the branch nodes below the current node.
}
\keyword{internal}

