% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{getChildData}
\alias{getChildData}
\title{Get the data of the child nodes below the current node.}
\usage{
getChildData(self, recursive = FALSE)
}
\arguments{
\item{self}{The node where to start.}

\item{recursive}{Should the function be called on all child nodes too?}
}
\value{
the data associated with child nodes.
}
\description{
Get the data of the child nodes below the current node.
}

