% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{getChildId}
\alias{getChildId}
\title{Get the ids of the child nodes below the current node.}
\usage{
getChildId(self, recursive = FALSE)
}
\arguments{
\item{self}{The node where to start.}

\item{recursive}{Should the function be called on all child nodes too?}
}
\value{
the ids associated with child nodes.
}
\description{
Get the ids of the child nodes below the current node.
}

