% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{getChildNodes}
\alias{getChildNodes}
\title{Get all the child nodes below the current node.}
\usage{
getChildNodes(self, recursive = FALSE)
}
\arguments{
\item{self}{The node where to start}

\item{recursive}{Should the function be called on all child nodes too?}
}
\value{
a list of child nodes.
}
\description{
Get all the child nodes below the current node.
}

