% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{getIterVal.generaltreeiter}
\alias{getIterVal.generaltreeiter}
\title{Function heavily inspired by iterators package.}
\usage{
\method{getIterVal}{generaltreeiter}(obj, plus = 0L, check = TRUE, ...)
}
\description{
Function heavily inspired by iterators package.
}
\keyword{internal}

