% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterators.R
\name{getIterVal}
\alias{getIterVal}
\title{Function heavily inspired by iterators package.}
\usage{
getIterVal(obj, plus, ...)
}
\description{
Function heavily inspired by iterators package.
}
\keyword{internal}

