% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{getLeafsData}
\alias{getLeafsData}
\title{Get the data of the leafs nodes below the current node.}
\usage{
getLeafsData(self, private, key)
}
\arguments{
\item{self}{The node where to start.}

\item{private}{the private members of the GeneralTree.}

\item{key}{The key for which the leafs should be returned}
}
\value{
the data associated with leafs nodes.
}
\description{
Get the data of the leafs nodes below the current node.
}
\keyword{internal}

