% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{getLeafsKeys}
\alias{getLeafsKeys}
\title{Get the keys of the leafs nodes below the current node.}
\usage{
getLeafsKeys(self, private, key)
}
\arguments{
\item{self}{The node where to start.}

\item{private}{the private members of the GeneralTree.}

\item{key}{The key for which the leafs should be returned}
}
\value{
the keys associated with leafs nodes.
}
\description{
Get the keys of the leafs nodes below the current node.
}
\keyword{internal}

