% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{getSiblingsInfo}
\alias{getSiblingsInfo}
\title{Get the keys of the siblings nodes below the current node.}
\usage{
getSiblingsInfo(self, private, key, what = c("key", "data"))
}
\arguments{
\item{self}{The node where to start.}

\item{private}{the private members of the GeneralTree.}
}
\value{
the keys associated with siblings nodes.
}
\description{
Get the keys of the siblings nodes below the current node.
}
\keyword{internal}

