% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{have_child}
\alias{have_child}
\title{Tell whether the current node has childeren.}
\usage{
have_child(self, private)
}
\arguments{
\item{self}{The node that should be inspected.}

\item{private}{The private members of a node.}
}
\value{
\code{TRUE} when the node has children.
}
\description{
Tell whether the current node has childeren.
}
\keyword{internal}

