% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{have_parent}
\alias{have_parent}
\title{Returns true when the node has a parent.}
\usage{
have_parent(self, private)
}
\arguments{
\item{self}{The node that should be inspected.}

\item{private}{The private members of a node.}
}
\value{
\code{TRUE} when the node has a parent.
}
\description{
Returns true when the node has a parent.
}
\keyword{internal}

