% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{have_siblings}
\alias{have_siblings}
\title{Tell whether the current node has siblings.}
\usage{
have_siblings(self, private)
}
\arguments{
\item{self}{The node that should be inspected.}

\item{private}{The private members of a node.}
}
\value{
\code{TRUE} when the node has siblings.
}
\description{
Tell whether the current node has siblings.
}
\keyword{internal}

