% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{id}
\alias{id}
\title{Return the id of the current node.}
\usage{
id(self, private)
}
\arguments{
\item{self}{The node that should be inspected.}

\item{private}{The private members of a node.}
}
\value{
the id of the node.
}
\description{
Return the id of the current node.
}
\keyword{internal}

