% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{initialize}
\alias{initialize}
\title{Initialize a General Tree object.}
\usage{
initialize(self, private, id, data)
}
\arguments{
\item{self}{the GeneralTree}

\item{private}{the private members of the GeneralTree.}

\item{id}{the id of the new node.}

\item{data}{the data of the new node.}
}
\description{
Initialize a General Tree object.
}
\keyword{internal}

