% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{is_last_sibling}
\alias{is_last_sibling}
\title{Tell whether the current node is the last sibling.}
\usage{
is_last_sibling(self, private)
}
\arguments{
\item{self}{The node that should be inspected.}

\item{private}{The private members of a node.}
}
\value{
\code{TRUE} when the node is the last sibling.
}
\description{
Tell whether the current node is the last sibling.
}
\keyword{internal}

