% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{is_root}
\alias{is_root}
\title{Tell whether the passed node is the root of the tree.}
\usage{
is_root(self, private)
}
\arguments{
\item{self}{The node that refers to the tree.}

\item{private}{The private members of a node.}
}
\value{
\code{TRUE} when the node is the root of the tree.
}
\description{
Tell whether the passed node is the root of the tree.
}
\keyword{internal}

