% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{isDiscovered}
\alias{isDiscovered}
\title{Informs whether the node has the discovered bit set.}
\usage{
isDiscovered(self, private)
}
\arguments{
\item{self}{The node where to start.}

\item{private}{The private members of a node.}
}
\value{
\code{TRUE} if the node has the discovered bit set.
}
\description{
Informs whether the node has the discovered bit set.
}
\keyword{internal}

