% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{isRootDiscovered}
\alias{isRootDiscovered}
\title{Returns whether the node has the is_root_discovered bit set.}
\usage{
isRootDiscovered(self, private)
}
\arguments{
\item{self}{The node where to start.}

\item{private}{The private members of a node.}
}
\value{
\code{TRUE} if the node has the is_root_discovered bit set.
}
\description{
Returns whether the node has the is_root_discovered bit set.
}
\keyword{internal}

