% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{isSingletonTree}
\alias{isSingletonTree}
\title{Informs whether the tree is a singleton tree.}
\usage{
isSingletonTree(self, private)
}
\arguments{
\item{self}{The node where to start.}

\item{private}{The private members of a node.}
}
\value{
\code{TRUE} if the node contains only a single element i.e. the root
 node.
}
\description{
Informs whether the tree is a singleton tree.
}
\keyword{internal}

