% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{iter.GeneralTree}
\alias{iter.GeneralTree}
\title{Internal function heavily inspired by iterators package.}
\usage{
iter.GeneralTree(obj, by = c("data"), checkFunc = function(...) TRUE,
  recycle = FALSE, ...)
}
\description{
Internal function heavily inspired by iterators package.
}
\keyword{internal}

