% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{iteratorImpl}
\alias{iteratorImpl}
\title{The implementation of the GeneralTree method iterator.}
\usage{
iteratorImpl(self, private)
}
\arguments{
\item{self}{the GeneralTree}

\item{private}{the private members of the GeneralTree.}
}
\value{
An iterator for the general tree.
}
\description{
The implementation of the GeneralTree method iterator.
}

