% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{left_child}
\alias{left_child}
\title{Return the left child of the node.}
\usage{
left_child(self, private)
}
\arguments{
\item{self}{The node that should be inspected.}

\item{private}{The private members of a node.}
}
\value{
The left child of the node or \code{NULL} otherwise.
}
\description{
Return the left child of the node.
}
\keyword{internal}

