% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{nextElem.GeneralTree}
\alias{nextElem.GeneralTree}
\title{Internal function heavily inspired by iterators package.}
\usage{
nextElem.GeneralTree(obj, ...)
}
\description{
Internal function heavily inspired by iterators package.
}
\keyword{internal}

