% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{nextElemWorkerImpl}
\alias{nextElemWorkerImpl}
\title{The implementation of the GeneralTree method nextElem.}
\usage{
nextElemWorkerImpl(self, private, set_discover = TRUE, include_root = TRUE)
}
\arguments{
\item{self}{the GeneralTree}

\item{private}{the private members of the GeneralTree.}

\item{set_discover}{Whether the discover flag should be set when seeking a
next element.}

\item{include_root}{Should the root be included in the discovery?}
}
\value{
the next element in the tree that has the discover flag not set
        searched in a depth first search.
}
\description{
The implementation of the GeneralTree method nextElem.
}
\keyword{internal}

