% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{nodeInfoToString}
\alias{nodeInfoToString}
\title{Convert a node to string.}
\usage{
nodeInfoToString(self, what = c("id", "data"))
}
\arguments{
\item{self}{the GeneralTree node that should be converted to a string.}

\item{what}{what should be converted to a string.}
}
\value{
A string object that represents the node.
}
\description{
Convert a node to string.
}
\keyword{internal}

