% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{parent}
\alias{parent}
\title{Get the parent of a node.}
\usage{
parent(self, private)
}
\arguments{
\item{self}{The node that refers to the tree.}

\item{private}{The private members of a node.}
}
\value{
the parent of the node.
}
\description{
Get the parent of a node.
}
\keyword{internal}

