% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.GeneralTree}
\alias{plot.GeneralTree}
\title{Plot a GeneralTree object.}
\usage{
\method{plot}{GeneralTree}(x, ...)
}
\arguments{
\item{x}{tree to plot.}

\item{...}{arguments passed to underlying functions.}
}
\description{
Plot a GeneralTree object.
}

