% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.GeneralTree}
\alias{print.GeneralTree}
\title{Print a GeneralTree object.}
\usage{
\method{print}{GeneralTree}(x, ...)
}
\arguments{
\item{x}{tree to print.}

\item{...}{arguments passed to underlying functions.}
}
\description{
Print a GeneralTree object.
}

