% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{resetDiscoveredOnBranch}
\alias{resetDiscoveredOnBranch}
\title{Reset the discover bit of all the nodes on the branch to FALSE.}
\usage{
resetDiscoveredOnBranch(self, private)
}
\arguments{
\item{self}{the GeneralTree}

\item{private}{the private members of the GeneralTree.}
}
\description{
Reset the discover bit of all the nodes on the branch to FALSE.
}
\keyword{internal}

