% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{root}
\alias{root}
\title{Return the root of the node.}
\usage{
root(self, private)
}
\arguments{
\item{self}{The node that should be inspected.}

\item{private}{The private members of a node.}
}
\value{
The root node.
}
\description{
Return the root of the node.
}
\keyword{internal}

