% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{searchNode}
\alias{searchNode}
\title{Search for an id in starting at a point in the tree and return the node
matching the id.}
\usage{
searchNode(self, id)
}
\arguments{
\item{self}{the node where to start searching.}

\item{id}{the id to look for.}
}
\value{
The data associated with an id.
}
\description{
Search for an id in starting at a point in the tree and return the node
matching the id.
}

