% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{<-.GeneralTree}
\alias{<-.GeneralTree}
\title{Deep clone a General Tree.}
\usage{
.GeneralTree(x) <- value
}
\arguments{
\item{x}{The general tree that should be deep cloned.}
}
\value{
a clone of the tree.
}
\description{
Deep clone a General Tree.
}

