% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{setDiscovered}
\alias{setDiscovered}
\title{Set the discovered bit to a certain state.}
\usage{
setDiscovered(self, private, is_discovered)
}
\arguments{
\item{self}{the GeneralTree}

\item{private}{the private members of the GeneralTree.}

\item{is_discovered}{the state to which the discover bit should be set.}
}
\description{
Set the discovered bit to a certain state.
}
\keyword{internal}

