% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{setRootDiscovered}
\alias{setRootDiscovered}
\title{Set the root discovered bit to a certain state.}
\usage{
setRootDiscovered(self, private, is_root_discovered)
}
\arguments{
\item{self}{the GeneralTree}

\item{private}{the private members of the GeneralTree.}

\item{is_root_discovered}{the state to which the bit should be set.}
}
\value{
An iterator for the general tree.
}
\description{
Set the root discovered bit to a certain state.
}
\keyword{internal}

