% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{siblings}
\alias{siblings}
\title{Return the private siblings of the node.}
\usage{
siblings(self, private)
}
\arguments{
\item{self}{The node that should be inspected.}

\item{private}{The private members of a node.}
}
\value{
A list that contains the private siblings of the node.
}
\description{
Return the private siblings of the node.
}
\keyword{internal}

