% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{toString}
\alias{toString}
\title{Convert a branch to a string.}
\usage{
toString(self, what = c("id", "data"), string_prepend = "")
}
\arguments{
\item{self}{the GeneralTree node from where the branch should be
converted to a string.}

\item{what}{what should be converted to a string.}

\item{string_prepend}{which string should be prepended to the string.}
}
\value{
A string object that represents the node.
}
\description{
Convert a branch to a string.
}
\keyword{internal}

