% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{travelUp}
\alias{travelUp}
\title{Travel one level up in the tree.}
\usage{
travelUp(self, private)
}
\arguments{
\item{self}{The point in the tree where the sibling should be added.}

\item{private}{The private part of the tree.}
}
\value{
invisible the new node that was created.
}
\description{
Travel one level up in the tree.
}
\keyword{internal}

