% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{treeDepth}
\alias{treeDepth}
\title{Get the depth of a tree.}
\usage{
treeDepth(self, private)
}
\arguments{
\item{self}{The node that refers to the tree.}

\item{private}{The private members of a node.}
}
\value{
A numeric value that indicates the number of layers in the tree.
}
\description{
Get the depth of a tree.
}
\keyword{internal}

