% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casting.R
\name{as.data.frame.GeneralTree}
\alias{as.data.frame.GeneralTree}
\title{Convert a GeneralTree to a data frame.}
\usage{
\method{as.data.frame}{GeneralTree}(x)
}
\description{
Convert a GeneralTree to a data frame.
}

