% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{deleteId}
\alias{deleteId}
\title{Delete a node with a given id.}
\usage{
deleteId(self, id)
}
\arguments{
\item{self}{The reference to the tree where the id should be searched.}

\item{id}{The id that should be deleted.}
}
\description{
Delete a node with a given id.
}

