% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\docType{class}
\name{GeneralTree}
\alias{GeneralTree}
\title{GeneralTree is a R6 implementation of a tree that can have multiple
childeren per parent.}
\format{\code{\link{R6Class}} object.}
\usage{
GeneralTree
}
\value{
Object of \code{\link{R6Class}} with methods for creating a general
tree.
}
\description{
GeneralTree is a R6 implementation of a tree that can have multiple
childeren per parent.
}
\section{Methods}{

\describe{
 \item{\code{addNode(parent_id, id, data)}}{Add a new node to the tree. The
 new node will be a child of parent_id and have an id and data.}
}
}
\examples{
root <- GeneralTree$new(0, 'root')
root$addNode(0, 1, 'child.0.1')
root$addNode(0, 2, 'child.0.2')
root$addNode(0, 3, 'child.0.3')
root$addNode(3, 4, 'child.3.4')
root$search(4)
}
\keyword{datasets}

