% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneralTree.R
\name{getBranchData}
\alias{getBranchData}
\title{Get the data of the branch nodes below the current node.}
\usage{
getBranchData(self, private, key)
}
\arguments{
\item{self}{The node where to start.}

\item{private}{the private members of the GeneralTree.}

\item{key}{The key for which the branch should be returned}
}
\value{
the data associated with branch nodes.
}
\description{
Get the data of the branch nodes below the current node.
}
\keyword{internal}

