% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterators.R
\name{getIterVal.GeneralTreeIter}
\alias{getIterVal.GeneralTreeIter}
\title{Function heavily inspired by iterators package.}
\usage{
\method{getIterVal}{GeneralTreeIter}(obj, plus = 0L, check = TRUE, ...)
}
\description{
Function heavily inspired by iterators package.
}
\keyword{internal}

