% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.GeneralTree}
\alias{print.GeneralTree}
\title{Print a GeneralTree object.}
\arguments{
\item{x}{General tree object.}

\item{...}{limits will set the maximum number of nodes to print.}
}
\description{
Print a GeneralTree object.
}

